pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "tests", "src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=src/simple_resume",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=85"
]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for component interaction",
    "business: Business logic and user story tests",
    "slow: Tests that take longer to run",
    "network: Tests that require network access",
    "tdd: Test-driven development tests"
]

[project]
name = "simple-resume"
version = "0.1.0"
description = "Generate PDF resumes from YAML data using HTML templates"
authors = [
    {name = "Alex Thola", email = "alexthola@gmail.com"}
]
maintainers = [
    {name = "Alex Thola", email = "alexthola@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
repository = "https://github.com/athola/simple-resume"
requires-python = ">=3.10"
dependencies = [
    "markdown>=3.7.0",
    "oyaml>=1.0",
    "weasyprint>=66.0",
    "palettable>=3.3.3",
]

[project.optional-dependencies]
utils = [
    "ruff>=0.14.0",
    "mypy>=1.18.0",
    "ty>=0.0.1a23",
    "pytype>=2024.10.11",
    "types-markdown>=3.9.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
simple-resume = "simple_resume.cli:main"

# Palette utilities
palette-snapshot = "simple_resume.palettes.cli:snapshot"
palette-list = "simple_resume.palettes.cli:list"
generate-random-palette-demo = "simple_resume.cli_random_palette_demo:main"

[tool.hatch.build]
include = [
    "src/simple_resume/templates/**/*",
    "src/simple_resume/static/**/*",
    "src/simple_resume/palettes/**/*",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "blacken-docs>=1.20.0",
    "interrogate>=1.7.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "beautifulsoup4>=4.12.3",
    "ty>=0.0.1a23",
    "safety>=3.6.2",
    "pdf2image>=1.17.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/simple_resume"]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [".venv", ".uv-cache", "src/simple_resume/static", "src/simple_resume/templates"]

[tool.ruff.lint]
# Rules: https://docs.astral.sh/ruff/rules/
# Using a subset of ruff-recommended rules plus some extras
# For a more aggressive linting, use:
# select = [
#   "ALL",
# ]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "S",   # flake8-bandit
  "PL",   # All Pylint rules (complete coverage)
  "D",   # pydocstyle
]
extend-ignore = [
    "E203",    # whitespace before ':' (conflicts with black/ruff format)
    "D203",    # incorrect-blank-line-before-class (incompatible with D211)
    "D213",    # multi-line-summary-second-line (incompatible with D212)
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = [
    "PLR2004",  # Allow magic numbers in tests (e.g., test data)
    "PLR0913",  # Allow more arguments in test functions (many fixtures)
    "PT009",    # Allow unittest-style assertions in pytest
    "PT011",    # Allow pytest.raises without match
    "RUF001",   # Allow unicode characters in test strings
    "RUF002",   # Allow unicode characters in docstrings
    "RUF003",   # Allow unicode characters in comments
    "S101",     # Allow assert statements in tests
    "D100",     # Relax docstring requirements for test modules
    "D101",     # Relax docstring requirements for test classes
    "D102",     # Relax docstring requirements for test methods
    "D103",     # Relax docstring requirements for test functions
    "D107",     # Relax docstring requirements for test __init__
    "D202",     # Allow blank lines after docstrings in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
mypy_path = "./src"
explicit_package_bases = true
namespace_packages = true
exclude = ['^\.venv', '^\.uv-cache']

[[tool.mypy.overrides]]
module = "oyaml"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "weasyprint"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false
